feat: implementação da funcionalidade de excluir imóvel

+++ src/main/java/br/com/invistaix/InvistaIX/service/ImovelService.java
@@ -8,7 +8,9 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import br.com.invistaix.InvistaIX.model.EnderecoModel;
 import br.com.invistaix.InvistaIX.model.ImovelModel;
+import br.com.invistaix.InvistaIX.repository.EnderecoRepository;
 import br.com.invistaix.InvistaIX.repository.ImovelRepository;
 import jakarta.transaction.Transactional;
 
@@ -17,6 +19,9 @@ public class ImovelService {
 
     @Autowired
     private ImovelRepository imovelRepository;
+    
+    @Autowired 
+    private EnderecoRepository enderecoRepository;
 
     public ImovelModel salvarImovel(ImovelModel imovel) {
         try {
@@ -106,7 +111,16 @@ public class ImovelService {
     		if (idImovel == null || idImovel <= 0) {
                 throw new IllegalArgumentException("ID do imóvel inválido.");
             }
+    		ImovelModel imovel = imovelRepository.findById(idImovel).orElse(null);
+    		if (imovel == null) {
+    			throw new IllegalArgumentException("Imóvel não encontrado");
+    		}
+    		EnderecoModel endereco = enderecoRepository.findById(imovel.getEndereco().getId()).orElse(null);
+    		if (endereco == null) {
+    			throw new IllegalArgumentException("Endereco não encontrado");
+    		}
     		imovelRepository.deleteById(idImovel);
+    		enderecoRepository.deleteById(endereco.getId());
     		return "Imóvel apagado com sucesso";
     	} catch (Exception ex) {
     		throw new RuntimeException("Erro ao apagar o imóvel: " + ex.getMessage(), ex);
